refactor(playground): split code analyzer checks
